+++ R/feistest.R
@@ -418,10 +418,12 @@ feistest <- function(model = NA, robust = FALSE, type = c("all", "art1", "art2",
 #'
 #' @examples
 #' data("mwp", package = "feisr")
+#' \dontrun{
 #' feis.mod <- feis(lnw ~ marry + enrol | year,
 #'                  data = mwp, id = "id", robust = TRUE)
-#' bsht <- bsfeistest(feis.mod, type = "bs1", rep = 5, seed = 1234, prog = FALSE)
+#' bsht <- bsfeistest(feis.mod, type = "bs1", rep = 100, seed = 1234)
 #' summary(bsht)
+#' }
 #'
 #' @export
 #'