+++ R/utility_functions.R
@@ -114,12 +114,13 @@ hatm <- function(y, x, weights = NULL, checkcol = TRUE, tol = .Machine$double.ep
 #'  \code{slopes} as variable(s). Otherwise must be excluded.
 #' @param slopes a \code{data.frame}, \code{matrix}, or \code{vector} of slopes to be used for detrending,
 #'  not containing an intercept. Optionally, a \code{character} vector of the names of slope variables
-#'  in \code{data}.
+#'  in \code{data}. For pure de-meaning please use "slopes = 1".
 #' @param id a \code{vector} of a unique group / person identifier. Optionally, a \code{character}
 #'  of the name of the unique group / person identifier in \code{data}. For overall detrending,
 #'  use \code{"id = 1"}.
-#' @param intercept logical. If \code{TRUE} the slopes will contain and an individual
+#' @param intercept logical. If \code{TRUE} the slopes will contain an individual
 #'  intercept (default is \code{TRUE}). For \code{"id = 1"}, this is an overall intercept.
+#'  Ignored if "slopes = 1".
 #' @param na.action character, either \code{na.exclude} (default) or \code{na.omit} indicates the use
 #'  of \code{NA}s. \code{na.exclude} passes \code{NA}s through to the output (same length as input).
 #'  \code{na.omit} drops \code{NA} rows (list-wise).
@@ -163,7 +164,12 @@ detrend <- function(data, slopes, id = NULL, intercept = TRUE,
     stop(paste("id is missing. For overall detrend, use 'id = 1'"))
   }
   if(is.null(slopes)){
-    stop(paste("slopes is missing with no default. Misspelled?"))
+    stop(paste("Argument slopes missing. Misspelled?"))
+  }
+  if(length(slopes) == 1 && slopes == 1){
+    dem <- TRUE
+  }else{
+    dem <- FALSE
   }
 
   ov <- FALSE
@@ -195,7 +201,7 @@ detrend <- function(data, slopes, id = NULL, intercept = TRUE,
   }
 
   # All slopes available
-  if(!is.character(slopes)){
+  if(!is.character(slopes) && !dem){
     if(is.vector(slopes)){
       ls <- length(slopes)
     }else{
@@ -204,6 +210,10 @@ detrend <- function(data, slopes, id = NULL, intercept = TRUE,
     if(ld != ls){
       stop(paste("slopes must have same length as data"))
     }
+  }
+
+  if(dem){
+    slopes <- rep(1, ld)
   }
 
   if(is.character(slopes)){
@@ -236,10 +246,15 @@ detrend <- function(data, slopes, id = NULL, intercept = TRUE,
   id <- c(id)[cp]
 
   # Make model matrix
-  fmz <- formula(paste0("~", paste0(colnames(slopes), collapse = "+")))
+  if(dem){
+    fmz <- formula(paste0("~", "1"))
+  }else{
+    fmz <- formula(paste0("~", paste0(colnames(slopes), collapse = "+")))
+  }
+
   Z <- model.matrix(fmz, slopes)
   # instead of update fm -1 use [,-1] (to avoid contrasts for intercept)
-  if(intercept == FALSE){
+  if(intercept == FALSE && slopes != 1){
     Z <- Z[, -1, drop = FALSE]
   }